% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_PDBPainter.R
\name{fct_get_chain_df}
\alias{fct_get_chain_df}
\title{Get a data frame of PDB chains.}
\usage{
fct_get_chain_df(pdb_code)
}
\arguments{
\item{pdb_code}{A string representing the PDB code (e.g., "7cid").}
}
\value{
A grouped `data.frame` with PDB chain information.
}
\description{
Code for Painting the PDBs with LogFC data.
}
