% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_PDBPainter.R
\name{fct_prepare_chainset}
\alias{fct_prepare_chainset}
\title{Combine and clean data to create a final color set for the viewer.}
\usage{
fct_prepare_chainset(joined_df)
}
\arguments{
\item{joined_df}{A data frame from [fct_join_logfc_chains()].}
}
\value{
A `data.frame` with 'sele' and 'colorValue' columns.
}
\description{
Combine and clean data to create a final color set for the viewer.
}
