% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_PDBPainter.R
\name{fct_range_logfc}
\alias{fct_range_logfc}
\title{Normalise logFC values for B-factor coloring}
\usage{
fct_range_logfc(df, range = c(-2, 2), rescale = FALSE)
}
\arguments{
\item{df}{A data frame containing a `logfc` column.}

\item{range}{A numeric vector of length 2 giving the min/max limits.}

\item{rescale}{Logical, whether to rescale values into [0,1]. Default = FALSE.}
}
\value{
The input data frame with an added column `bfactor_value`.
}
\description{
This function clamps logFC values into a user-selected range
and rescales them (if desired) to keep values consistent across
structures.
}
