% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_insertUI_component.R
\name{insertUI_floating_panel}
\alias{insertUI_floating_panel}
\title{Insert Floating Panel}
\usage{
insertUI_floating_panel(panel_id, r)
}
\arguments{
\item{panel_id}{The module id for the floating panel.}

\item{r}{The global reactiveValues object containing \code{r$floatingPanel$inserted}.}
}
\description{
Helper to inject the floating selection panel into the document body.
}
