% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_insertUI_component.R
\name{insertUI_selection}
\alias{insertUI_selection}
\title{components UI selection}
\usage{
insertUI_selection(sequence, selection, gene_map)
}
\arguments{
\item{sequence}{protein sequence in string format.}

\item{selection}{ngl.js query to be transformed. Uses \code{selection_to_ngl()}}

\item{gene_map}{gene map selection.}
}
\description{
Function to transform ngl.js selection for UI component
}
\examples{
\dontrun{
insertUI_selection("1-3 OR <GHK>", "NGLSDFGHK")
}
}
