% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_selection_to_ngl.R
\name{selection_to_ngl}
\alias{selection_to_ngl}
\title{Selection to NGL query}
\usage{
selection_to_ngl(sequence, input, gene_map = NULL)
}
\arguments{
\item{sequence}{Protein sequence in string format}

\item{input}{NGL query}

\item{gene_map}{gene map selection.}
}
\description{
Transform sequence input to NGL query
}
\examples{
selection_to_ngl("ALAAGSDFG", "1-5 OR <DFG>")
}
